% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divergence_Estimate.R
\name{divEst}
\alias{divEst}
\title{Estimate vagrant insert proportion from diverged population data}
\usage{
divEst(dat)
}
\arguments{
\item{dat}{A data.frame with the following eleven columns in this order.
\describe{
  \item{pos}{A factor (or structure that can be coerced to a factor),
  giving the site IDs}
  \item{sample}{A factor (or structure that can be coerced to a factor),
  giving a unique name for each individual genotyped}
  \item{g1}{A numeric vector giving the allel count of allele 1.}
  \item{g2}{A numeric vector giving the allel count of allele 2.}
  \item{g3}{A numeric vector giving the allel count of allele 3.}
  \item{g4}{A numeric vector giving the allel count of allele 4.}
  \item{N}{A numeric vector giving the number of bp in the read data that did not map
   to the vargrant DNA reference in this individual.}
  \item{M}{A numeric vector giving the number of bp in the read data that did map
  to the vargrant DNA reference in this individual.}
  \item{pop}{A factor (or structure that can be coerced to a factor), of "A" and "B"
  denoting which population the individual belongs to.}
  \item{A}{A numeric vector giving the number of the major allele at this site
  in population A.}
  \item{B}{A numeric vector giving the number of the major allele at this site
  in population B.}
  }}
}
\value{
An (invisible) list of Ascores and Bscores corresponding to the
estimates per locus using A or B as the focal population.
}
\description{
Estimate vagrant insert proportion from diverged population data
}
\examples{
## Access one of the package's example data-sets (parrotFX or hopperFX)
data(hopperFX)

## plot and printout (by default) the results of running divEst on the grasshopper
divEst(hopperFX)
##
## save the estimates from each locus and population and locus
res1 <- divEst(hopperFX)
## Inspect the individual scores
str(res1)

}
