% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interceptPositionPlot.R
\name{interceptPositionPlot}
\alias{interceptPositionPlot}
\title{Intercept-position-plot for loci selected by \code{rainbowPlot()}}
\usage{
interceptPositionPlot(
  ff,
  highlightOutliers = F,
  main = "Intercept-position-plot",
  ...
)
}
\arguments{
\item{ff}{A \code{rainbowPlot()} fit.}

\item{highlightOutliers}{Logical. Whether or not to highlight sites with
outlier intercepts according to \code{boxplot.stats}. Default \code{F}}

\item{main}{Plot caption, default is "Intercept-position-plot"}

\item{...}{Arguments to be passed to \code{plot()}}
}
\value{
Returns invisibly a vector of positions where the intercept estimate
were outliers according to \code{boxplot.stats}
}
\description{
Intercept-position-plot for loci selected by \code{rainbowPlot()}
}
\examples{
 \dontrun{
 rPar <- rainbowPlot(parrotDF)
 toRemove <- interceptPositionPlot(rPar, highlightOutliers=T)
 parrotDF2 <- parrotDF[!parrotDF$Position \%in\% toRemove,]
 rainbowPlot(parrotDF2)
 }
}
