% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selAllPlot.R
\name{selAllPlot}
\alias{selAllPlot}
\title{Boxplot of alternate (vagrant) allele counts per sample}
\usage{
selAllPlot(dd, ff, ylim = c(0, 10), ylab = "Allele count", ...)
}
\arguments{
\item{dd}{A dataframe, such as used to run `rainbowPlot()`}

\item{ff}{(optional) a `rainbowPlot()` fit}

\item{...}{Arguments passed to `boxplot()`}
}
\description{
Boxplot of alternate (vagrant) allele counts per sample
}
\examples{
#'  \dontrun{
 humanFit <- rainbowPlot(humanDF)
 selAllPlot(humanDF) # overall alt allele counts
 selAllPlot(humanDF, humanFit) # alt allele count at loci selected for fit
 }
}
