% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interceptPositionPlot.R
\name{interceptPositionPlot}
\alias{interceptPositionPlot}
\title{Intercept-position-plot for loci seect by rainbowPlot()}
\usage{
interceptPositionPlot(ff, main = "Intercept-position-plot", ...)
}
\arguments{
\item{ff}{A `rainbowPlot()` fit.}

\item{main}{Plot caption, default is "Intercept-position-plot"}

\item{...}{Arguments to be passed to `plot()`}
}
\description{
Intercept-position-plot for loci seect by rainbowPlot()
}
\examples{
 \dontrun{
 humanFit <- rainbowPlot(humanDF)
 interceptPositionPlot(humanFit)
 }
}
